TASK#14: Refactoring Actions were moved into separated folders
