+++ project/src/app.tsx
@@ -9,11 +9,11 @@ import NotFound from './screens/not-found/not-found';
 import { AppRoute } from './constants';
 import PrivateRoute from './components/private-route/private-route';
 import { AsyncDispatch } from './types/action';
-import { checkAuthStatus } from './store/user/thunks';
+import { checkAuthStatus } from './store/user/user-thunks';
 import { connect, ConnectedProps } from 'react-redux';
 import { State } from './store/reducer';
 import React, { useEffect } from 'react';
-import { setAppInitialized } from './store/action';
+import { setAppInitialized } from './store/app/app-actions';
 import Sprite from './components/sprite/sprite';
 
 const mapStateToProps = (state: State) => ({
