--- project/src/store/app/type.ts
@@ -1,4 +1,0 @@
-export type AppState = {
-  genre: string;
-  appInitialized: boolean;
-};
