--- project/src/store/user/type.ts
@@ -1,7 +1,0 @@
-import { AuthStatus } from '../../constants';
-import { User } from '../../types/user';
-
-export type UserState = {
-  authStatus: AuthStatus,
-  user: User | null,
-};
