+++ project/src/store/user/user-thunks.ts
@@ -1,10 +1,10 @@
-import { setAuthStatus, setUserData } from '../action';
 import { dropToken, setToken } from '../../services/token';
 import { AuthStatus, BackendRoute } from '../../constants';
 import { SignInForm } from '../../types/sign-in-form';
 import { ServerUser } from '../../types/user';
 import { AsyncAction } from '../../types/action';
 import { adaptUserToClient } from '../../services/adaptor';
+import { setAuthStatus, setUserData } from './user-actions';
 
 const login = (signIn: SignInForm): AsyncAction =>
   async (dispatch, _getState, api): Promise<void> => {
