+++ project/src/types/action.ts
@@ -1,27 +1,14 @@
-import {
-  setAppInitialized,
-  setAuthStatus,
-  setGenre,
-  setUserData
-} from '../store/action';
 import { ThunkAction, ThunkDispatch } from '@reduxjs/toolkit';
 import { AxiosInstance } from 'axios';
-import { State } from '../store/reducer';
+import { AppActions } from '../store/app/app-actions';
 import { DataActions } from '../store/data/data-actions';
-
-export enum ActionType {
-  SetGenre = 'app/set-genre',
-  SetAppInitialized = 'app/set-app-initialized',
-  SetAuthStatus = 'user/set-auth-status',
-  SetUserData = 'user/set-user-data',
-}
+import { State } from '../store/reducer';
+import { UserActions } from '../store/user/user-actions';
 
 export type Actions =
-  | ReturnType<typeof setGenre>
-  | ReturnType<typeof setAuthStatus>
-  | ReturnType<typeof setUserData>
-  | ReturnType<typeof setAppInitialized>
-  | DataActions;
+  | DataActions
+  | UserActions
+  | AppActions;
 
 export type AsyncAction<R = Promise<void>> = ThunkAction<R, State, AxiosInstance, Actions>
 export type AsyncDispatch = ThunkDispatch<State, AxiosInstance, Actions>;
